Fix: Convert requiredFuel method return type to int

+++ src/main/java/prj/fleet/Aircraft.java
@@ -29,7 +29,7 @@ public abstract class Aircraft {
         return this.flightNumber;
     }
 
-    public double getRequiredFuel() {
+    public int getRequiredFuel() {
         return this.requiredFuel;
     }
